wikipedia-and-math/requirements.txt
langchain
openai
streamlit
python-dotenv
numexpr
wikipedia